Refined HumdrumFileStreamer.

+++ src/HumdrumFileStructure.cpp
@@ -1410,7 +1410,7 @@ int HumdrumFileStructure::getStrandCount(int spineindex) const {
 	if (spineindex < 0) {
 		return 0;
 	}
-	if (spineindex >= m_strand2d.size()) {
+	if (spineindex >= (int)m_strand2d.size()) {
 		return 0;
 	}
 	return (int)m_strand2d[spineindex].size();
